Add Procedure
#Procedure A: Phone Number Lookup 
The procedure takes a phone number as input and return a character string in an output variable. The value output is either the name of the subject who uses the phone or ‘Unrecognised’.

#Procedure B: Visit Data Archiving
The procedure takes a date as input. On execution, all visit records more than 21 days old are placed in an archive and the data is then deleted from the main database. The first time the procedure is executed, it creates a table containing the archived data. On subsequent executions, it adds new data to the existing archive table.
Finally, the procedure output to screen the full contents of the archive table.